+++ tests/cas04/02b_process_cas04_mth5.py
@@ -42,9 +42,9 @@ from aurora.config.config_creator import ConfigCreator
 from aurora.general_helper_functions import TEST_PATH
 from aurora.pipelines.process_mth5 import process_mth5
 
-from aurora.tf_kernel.dataset import KernelDataset
 from aurora.tf_kernel.run_summary import RunSummary
 from aurora.transfer_function.plot.comparison_plots import compare_two_z_files
+from aurora.transfer_function.kernel_dataset import KernelDataset
 from mth5.utils.helpers import initialize_mth5
