+++ tests/synthetic/test_stft_methods_agree.py
@@ -10,8 +10,8 @@ from aurora.pipelines.time_series_helpers import run_ts_to_stft
 from aurora.pipelines.time_series_helpers import run_ts_to_stft_scipy
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test1_h5
 from aurora.test_utils.synthetic.make_processing_configs import create_test_run_config
-from aurora.tf_kernel.dataset import KernelDataset
 from aurora.tf_kernel.run_summary import RunSummary
+from aurora.transfer_function.kernel_dataset import KernelDataset
 from mth5.mth5 import MTH5
